requirements.txt
pypdf2
pdfminer
pdfminer.six
pdfplumber
pdf2image
pillow
pytesseract
openai
pinecone-client
requests
os
json
re